Formatação

+++ src/utilidades.c
@@ -32,7 +32,7 @@ int procura_vaga (const void* lista, size_t tam, uint8_t tipo) {
 	if (tipo == INDIVIDUO || tipo == INDIVIDUO) {
 		individuo* p = (individuo*)lista;
 
-		for (int i = 0; i < tam; ++i) {
+		for (size_t i = 0; i < tam; ++i) {
 			if (p[i].matricula == NAO_ATIVO) {
 				return i;
 			}
@@ -42,7 +42,7 @@ int procura_vaga (const void* lista, size_t tam, uint8_t tipo) {
 	else if (tipo == DISCIPLINA || tipo == DISCIPLINA) {
 		disciplina* p = (disciplina*)lista;
 
-		for (int i = 0; i < tam; ++i) {
+		for (size_t i = 0; i < tam; ++i) {
 			if (p[i].codigo[0] == NAO_ATIVO) { 
 				return i;
 			}
